Reorganized the header files (issue #3).
The header files are all put in the ChimeraTK/ControlSystemAdapter directory,
so that they use the same path in the source and the install directory.

Three new header files are added:
 - ChimeraTK/ControlSystemAdapter.h
 - ChimeraTK/ControlSystemAdapterControlSystem.h
 - ChimeraTK/ControlSystemAdapterDevice.h

They can be used to include all headers or all headers, the headers needed for
the control-system side, or the headers needed for the device side.