Implement copy-on-send for ProcessArrays (issue #5).
When using send() on a ProcessArray, the array’s value is now copied by default.
A new sendDestructively() method has been added that allows for sending a value
without copying. This method is only available if the corresponding flag has
been set when creating the process array.

As a result of this change, the getConst(), getLastSent(),
getLastSentVersionNumber(), and isSwappable() methods have been removed and a
ProcessArray’s value cano now always be modified, even if it is a receiver.