Update necessary for changes in DeviceAccess: The TransferFuture now transports a data structure containing the version number.
To implement this, the internal lockfree queue of the ProcessArray has been changed to contain futures of pointers to the buffer. The buffer class is now derived from the TransferFuture::Data class, so the futures in the lockfree queue are compatible with the furure expected by the TransferFuture class.

Since only spsc_queues can store even shared_futures, the fullBufferQueue had to be changed into an spsc_queue. This implies a behavioural change:

When the queue is full and data has to be thrown away, the ProcessArray can no longer throw away the oldest data, since it cannot pop from the spsc_queue. To make sure that the latest data can be kept, an internal atomic triple buffer has been introduced, which holds the latest data when the queue is full. Thus, even in the case of a buffer overflow, readLatest() will still return the latest value.