Implement bidirectional process arrays (issue #2).
This commit adds the bidirectional process-array implementation and
makes it available through the PVManager and DevicePVManager.
It also makes a few minor modifications to the unidirectional process
array. In particular, the unidirectional process array now checks that
the version number used in a write operation is greater than the last
version number used. This should help to detect bugs in device
libraries that incorrectly deal with version numbers.