Lockfree PersistentDataStorage::updateValue
introduced a lock free queue with update Value as producer and
retrieveValue and wrieToFile as consumers.

Read access to the queue is protected by a mutex; this is resonable
since retrieveValue accesses the queue only during the starup phase. The
writer thread would not block on this mutex once the server is running.