leave initialization to user by default
concept of initialization was not very useful since underlying buffer
was allowed to be swapped out. So turn it off by default and leave it up
to user, whether he wants to initialize.
Instead, in MappedImage::setShape header then needs to be fully
initialized.
The change should give some performance improvement since it removes
unneccesary initialization